+++ .openshiftio/application.yaml
@@ -1,7 +1,7 @@
 apiVersion: v1
 kind: Template
 metadata:
-  name: spring-boot-configmap
+  name: configmap
   annotations:
     iconClass: icon-spring
     tags: spring-boot, configmap, java, microservice
@@ -54,7 +54,7 @@ objects:
 - apiVersion: v1
   kind: ImageStream
   metadata:
-    name: spring-boot-configmap
+    name: configmap
   spec: {}
 - apiVersion: v1
   kind: ImageStream
@@ -69,12 +69,12 @@ objects:
 - apiVersion: v1
   kind: BuildConfig
   metadata:
-    name: spring-boot-configmap
+    name: configmap
   spec:
     output:
       to:
         kind: ImageStreamTag
-        name: spring-boot-configmap:BOOSTER_VERSION
+        name: configmap:BOOSTER_VERSION
     postCommit: {}
     resources: {}
     source:
@@ -111,11 +111,11 @@ objects:
   kind: Service
   metadata:
     labels:
-      app: spring-boot-configmap
+      app: configmap
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
-    name: spring-boot-configmap
+      group: dev.snowdrop.example
+    name: configmap
   spec:
     ports:
     - name: http
@@ -123,25 +123,25 @@ objects:
       protocol: TCP
       targetPort: 8080
     selector:
-      app: spring-boot-configmap
+      app: configmap
       provider: snowdrop
-      group: io.openshift.booster
+      group: dev.snowdrop.example
 - apiVersion: v1
   kind: DeploymentConfig
   metadata:
     labels:
-      app: spring-boot-configmap
+      app: configmap
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
-    name: spring-boot-configmap
+      group: dev.snowdrop.example
+    name: configmap
   spec:
     replicas: 1
     revisionHistoryLimit: 2
     selector:
-      app: spring-boot-configmap
+      app: configmap
       provider: snowdrop
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     strategy:
       rollingParams:
         timeoutSeconds: 3600
@@ -149,10 +149,10 @@ objects:
     template:
       metadata:
         labels:
-          app: spring-boot-configmap
+          app: configmap
           provider: snowdrop
           version: "BOOSTER_VERSION"
-          group: io.openshift.booster
+          group: dev.snowdrop.example
       spec:
         containers:
         - env:
@@ -160,7 +160,7 @@ objects:
             valueFrom:
               fieldRef:
                 fieldPath: metadata.namespace
-          image: spring-boot-configmap:BOOSTER_VERSION
+          image: configmap:BOOSTER_VERSION
           imagePullPolicy: IfNotPresent
           name: spring-boot
           ports:
@@ -184,20 +184,20 @@ objects:
         - spring-boot
         from:
           kind: ImageStreamTag
-          name: spring-boot-configmap:BOOSTER_VERSION
+          name: configmap:BOOSTER_VERSION
       type: ImageChange
 - apiVersion: v1
   kind: Route
   metadata:
     labels:
-      app: spring-boot-configmap
+      app: configmap
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
-    name: spring-boot-configmap
+      group: dev.snowdrop.example
+    name: configmap
   spec:
     port:
       targetPort: 8080
     to:
       kind: Service
-      name: spring-boot-configmap
+      name: configmap
