+++ README.md
@@ -1,3 +1,3 @@
-[![CircleCI](https://circleci.com/gh/snowdrop/spring-boot-configmap-booster.svg?style=shield)](https://circleci.com/gh/snowdrop/spring-boot-configmap-booster)
+[![CircleCI](https://circleci.com/gh/snowdrop/configmap-example.svg?style=shield)](https://circleci.com/gh/snowdrop/configmap-example)
 
 https://appdev.openshift.io/docs/spring-boot-runtime.html#mission-configmap-spring-boot
