+++ src/main/java/dev/snowdrop/example/service/GreetingController.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.openshift.booster.service;
+package dev.snowdrop.example.service;
 
 import java.util.Objects;
