+++ src/main/resources/static/index.html
@@ -14,7 +14,7 @@
         <p class="lead">
             This mission showcases application configuration, using OpenShift <em>configmaps</em>. It
             demonstrates application and runtime configuration leveraging external configuration sources. Alongside
-            the basic means to set up a <em>configmap</em> and use with a specific runtime, this booster also
+            the basic means to set up a <em>configmap</em> and use with a specific runtime, this example also
             demonstrates how changes to the configuration can be applied to services already deployed in OpenShift.
         </p>
     </div>
